style: template을 메서드에서 props로 변경
