+++ src/App.template.ts
@@ -1,0 +1,10 @@
+const template = () => {
+  return `
+    <main id="notion-app">
+      <nav id="notion-app-sidebar"></nav>
+      <section id="notion-app-content"></section>
+    </main>
+  `
+}
+
+export default template
