+++ src/components/Editor/index.ts
@@ -1,14 +1,7 @@
 import Component from '~/core/component'
-import { IDocument, OnEditing } from '~/models/document'
-
-type State = {
-  document: IDocument | null
-}
-
-interface Props {
-  parentElement: HTMLElement
-  onEditing: OnEditing
-}
+import { OnEditing } from '~/models/document'
+import template from './template'
+import { Props, State } from './types'
 
 const initialState: State = {
   document: null,
@@ -17,22 +10,10 @@ const initialState: State = {
 export default class Editor extends Component<State> {
   private onEditing: OnEditing
   constructor({ parentElement, onEditing }: Props) {
-    super({ parentElement, initialState, tag: 'div' })
+    super({ parentElement, initialState, template, tag: 'div' })
     this.element.className = 'editor'
     this.onEditing = onEditing
     this.attachEventHandler('keyup', this.handleKeyup)
-  }
-
-  template(state: State): string {
-    const { document } = state
-    if (!document) {
-      return ''
-    }
-    const { title, content } = document
-    return `
-      <input name="title" placeholder="제목을 입력하세요." value="${title}" />
-      <textarea name="content" placeholder="내용을 입력하세요." >${content}</textarea>
-      `
   }
 
   private handleKeyup(e: Event) {
