+++ src/components/Modal/types.ts
@@ -1,0 +1,11 @@
+import { OnSubmit } from '~/models/document'
+
+export type State = {
+  parentNodeId?: number
+  isView: boolean
+}
+
+export interface Props {
+  parentElement: HTMLElement
+  onSubmit: OnSubmit
+}
