+++ src/components/Sidebar/LastNode/types.ts
@@ -1,0 +1,9 @@
+export type State = {
+  isParentSpread: boolean
+  depth: number
+}
+
+export interface Props {
+  parentElement: HTMLElement
+  initialState: State
+}
